requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14,<3.16
Pillow>=10.0
drf-spectacular>=0.27,<0.28
django-filter>=24.2
psycopg2-binary>=2.9
python-dotenv>=1.0
PyJWT>=2.9
djangorestframework-simplejwt>=5.3
